update bids_tools and stitching script to current status quo
